refactor: move api event and flutter subscriber to `event::api`
